Don't call vc-state unless the file exists
If a buffer is visiting a file whose parent directory has been
renamed, vc-state throws an error because (buffer-file-path) no longer
points to an existing file.

Ensure that we never call vc-state when the file doesn't exist. This
issue can occur with diff-hl-flydiff, or when using
diff-hl-magit-post-refresh as documented in the readme.

This change causes diff-hl to stop highlighting in the current buffer,
which is the best we can do when we can't find the file.