https://github.com/dgutov/diff-hl/pull/147#issuecomment-735318376
I think it should be easy enough to extract a helper function from that package into the diff-hl package proper and use it from both diff-hl-flydiff-mode and the new feature.

As the first approximation, you can even autoload diff-hl-flydiff-buffer-with-head and use it directly (I can take a look later at whether further reorganization is really necessary).

Also, instead of calling diff-hl-diff-goto-hunk itself, you can create a specialized version of that function just for use in diff-hl-show-hunk that would call diff-hl-flydiff-buffer-with-head instead of vc-diff-internal. It wouldn't need to call diff-hl-remove-overlays, and it would use a direct control flow without vc-exec-after because diff-hl-flydiff-buffer-with-head uses 'noasync.