Added diff-hl-show-hunk-highlight-clicked-line.
Since `diff-hl-show-hunk-ensure-visible` moves the buffer point to the beggining of the hunk, I think it has no sense to highlight the line where the popup was invoked.

I have added diff-hl-show-hunk-highlight-clicked-line variable, to control if the line should be highlighted or not. Also, I made the `line` parameter of the backend function optional.

Currently, I think `ensure-visible` and `highlight-clicked-line` are a little bit messy, but I don't kown how to make them clearer:

- `ensure-visible` controls two behaviours: (1) It scrolls the buffer to show the entire hunk and (2) move the point to the beggining of the hunk. Two variables, maybe?

- If `ensure-visible` is `t` then `highlight-clicked-line` should be `nil`. Only one variable, maybe?