Navigation between hunks revamped, and some other issues
> There is a bug I'm working on: when going to previous hunk from the first hunk, No further hunks found appears in minibuffer instead of There is no previous change.

Solved.

> You can use the :overline property instead. And I'd also like to see no empty line above the header.

There is a problem in console (not GUI): console doesn't not support overline, so I simulated it with an empty line with underline. I added a `display-graphic-p` test, hope it will have no problem with the GUI and the console is used at the same time (via `emacsclient`)

> And when it's non-nil, I think it would be natural to have diff-hl-show-hunk-inline-popup-hide-hunk set to nil.

Currently, I prefer `-inline-popup-hide-hunk` set to `t` and `-smart-lines` set to `t` :) For me is like a wayback machine. I think I have set the default values to your preferences, tell me if not.

> You're right about plain empty popup being a problem, though. Maybe it should show something, maybe it should display some clarification text,

I have added the `diff-hl-show-hunk--no-lines-removed-message` constant, that is shown when `-smart-lines` set to `t` and the hunk has only additions.

> Will you remove the scrollable-quick-peek backend?

I forgot it. Removed.