Fix compiler warnings involving macro
The macro `diff-hl-with-diff-switches' is define and used in
"diff-hl.el".  Additionally it is used in "diff-hl-flydiff.el".

This macro uses variables from the `vc' package without requiring it
up front.  Until now this was dealt with using top-level variable
declarations in "diff-hl.el".  That is not enough because the
expansion of that macro, in another file, does not benefit from
these declarations.

  diff-hl-flydiff.el:47:8: Warning: Unused lexical variable ‘vc-svn-diff-switches’
  diff-hl-flydiff.el:47:8: Warning: Unused lexical variable ‘vc-fossil-diff-switches’
  diff-hl-flydiff.el:47:8: Warning: Unused lexical variable ‘vc-jj-diff-switches’

So we have to add the variable declarations to the macro expansion
themselves.