Support using reference revision as project-local
* diff-hl.el (diff-hl-mode): Set and clear buffer-local value of the reference
revision, if the buffur belongs to a project with the custom reference revision,
set by diff-hl-set-reference-rev-in-project.
(diff-hl-reference-revision): Make it safe if string or nil.
(diff-hl-set-reference-rev): Add resetting options when called with the prefix
argument.
(diff-hl-set-reference-rev-in-project): Add new command.
(diff-hl-set-reference-rev-in-project-internal): New internal function.
(diff-hl-reset-reference-rev): Add logic for calling with the prefix argument
and turn off amend-mode.
(diff-hl-reset-reference-rev-in-project): Add new command.
(diff-hl--project-root): Add new internal function.