Fix compiler warnings about project symbols properly
It is not appropriate to declare that a function exists, when we
very well know that it does in fact not exist (anymore).  That was
the approach taken in #254.  If we revert that, we are back to:

  diff-hl.el:1677:19: Warning: the function ‘project-buffers’ is not known to be define
  diff-hl.el:1668:9: Warning: the function ‘project-roots’ is not known to be defined.
  diff-hl.el:1654:17: Warning: the function ‘project-name’ is not known to be defined.

Once we add:

  (eval-when-compile (require 'project))

we are left with

  diff-hl.el:414:16: Warning: ‘revert-buffer-in-progress-p’ is an
  obsolete variable (as of 31.1); use ‘revert-buffer-in-progress’ instead.

which we deal with using `static-if'.  Unfortunately that macro was
only added in Emacs 30.1.  To get our hands on it earlier, we have
two options. Either we start depending on the Compat package, which
provides this macro, or we define it ourselves.

I went with the latter approach, assuming that you wouldn't want to
add an additional dependency just for this.  But surely it is not okay
to (re-)define a macro that is usually defined in (later versions of)
Emacs?!  Glad you asked!  The implementation in "subr.el" is
prefixed with:

  ;; Note: `static-if' can be copied into a package to enable it to be
  ;; used in Emacsen older than Emacs 30.1.  If the package is used in
  ;; very old Emacsen or XEmacs (in which `eval' takes exactly one
  ;; argument) the copy will need amending.